Fix the exception of ArrayIndexOutOfBoundsException when parsing the setting file, and add catch for ArrayIndexOutOfBoundsException to ensure that the backup file can be parsed
After inserting or replacing settings_global.xml, settings_system.xml, and settings_ssaid.xml, an ArrayIndexOutOfBoundsException exception will be thrown. Google has implemented backup and restoration for these files, but it has not caught the ArrayIndexOutOfBoundsException exception. When the source file is damaged, the logic of parsing the backup file will not go, resulting in JE Crash on the system_server, the phone restarts frequently, enters recovery, and cannot boot normally.

Test: as follows
     - Pull settings_global.xml or settings_system.xml or settings_ssaid.xml
     - Insert garbled characters into multiple positions in the file or replace part of the content with garbled characters
     - Push settings_global.xml or settings_system.xml or settings_ssaid.xml to the corresponding directory
     - Restart the phone and observe the power-on status of the phone

Change-Id: If7d4b1af01440329f4c93947410fc5d8856e055c
Signed-off-by: zouning <zouning@xiaomi.corp-partner.google.com>